refactor: entities/comment 제거 및 import 경로 업데이트
