--- src/entities/comment/index.ts
@@ -1,20 +1,0 @@
-export type {
-  Comment,
-  CommentCountResponse,
-  CreateCommentRequest,
-  SyncMeetingRequest,
-  UpdateCommentRequest,
-} from './api/comment.api';
-export { commentApi } from './api/comment.api';
-export {
-  commentKeys,
-  useCommentCount,
-  useComments,
-  useCreateComment,
-  useDeleteComment,
-  useLikeComment,
-  useSyncCreateMeeting,
-  useUpdateComment,
-} from './model/comment.queries';
-export type { CommentInputProps } from './ui/comment-input';
-export { CommentInput } from './ui/comment-input';
