--- src/entities/comment/ui/comment-input/index.ts
@@ -1,2 +1,0 @@
-export { CommentInput } from './comment-input';
-export type { CommentInputProps } from './comment-input.types';
