--- src/entities/comment/ui/comment-item/index.ts
@@ -1,2 +1,0 @@
-export { CommentItem } from './comment-item';
-export type { CommentItemData, CommentItemProps } from './comment-item.types';
