+++ src/widgets/sosotalk/ui/sosotalk-comment-section/sosotalk-comment-section.types.ts
@@ -1,24 +1,4 @@
-export interface SosoTalkCommentItemData {
-  id: string;
-  authorName: string;
-  authorImageUrl?: string;
-  createdAt: string;
-  relativeTime?: string;
-  content: string;
-  isAuthorComment?: boolean;
-  isEditing?: boolean;
-  editValue?: string;
-  isEditPending?: boolean;
-  onEditClick?: () => void;
-  onDeleteClick?: () => void;
-  onEditValueChange?: (value: string) => void;
-  onEditSubmit?: () => void;
-  onEditCancel?: () => void;
-}
-
-export interface SosoTalkCommentItemProps extends Omit<SosoTalkCommentItemData, 'id'> {
-  className?: string;
-}
+import type { SosoTalkCommentItemData } from '@/entities/sosotalk-comment';
 
 export interface SosoTalkCommentSectionProps {
   comments: SosoTalkCommentItemData[];
