+++ src/widgets/sosotalk/ui/sosotalk-post-detail/model/sosotalk-post-detail-page.utils.ts
@@ -2,6 +2,7 @@ import { format } from 'date-fns';
 import { ko } from 'date-fns/locale';
 
 import type { GetSosoTalkPostDetailResponse } from '@/entities/post';
+import type { SosoTalkCommentItemData } from '@/entities/sosotalk-comment';
 import type { Comment } from '@/shared/types/generated-client/models/Comment';
 import type { SosoTalkCommentItemData } from '../../sosotalk-comment-section/sosotalk-comment-section.types';
